+++ apps/cold-api/src/platform/modules/resources/integrations/integrations.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
-import { CacheService, ColdCacheModule, ColdRabbitModule, ColdRabbitService, PrismaModule } from '@coldpbc/nest';
+import { ColdCacheModule, ColdRabbitModule, MqttModule, PrismaModule } from '@coldpbc/nest';
 import { IntegrationsController } from './integrations.controller';
 import { IntegrationsService } from './integrations.service';
-import { OrganizationLocationsModule } from '../organization_locations/organization_locations.module';
+import { LocationsModule } from '../organizations/locations/locations.module';
+import { BroadcastEventService } from '../../../utilities/events/broadcast.event.service';
 
 @Module({
-  imports: [PrismaModule, ColdCacheModule, ColdRabbitModule.forFeature(), OrganizationLocationsModule],
+  imports: [PrismaModule, ColdCacheModule, ColdRabbitModule.forFeature(), LocationsModule, MqttModule],
   controllers: [IntegrationsController],
-  providers: [IntegrationsService, CacheService, ColdRabbitService],
-  exports: [IntegrationsService, ColdRabbitService],
+  providers: [IntegrationsService, BroadcastEventService],
+  exports: [IntegrationsService],
 })
 export class IntegrationsModule {}