+++ apps/cold-api/src/platform/modules/resources/service_definitions/service_definitions.module.ts
@@ -1,12 +1,13 @@
 import { Module } from '@nestjs/common';
-import { CacheService, ColdCacheModule, ColdRabbitModule, ColdRabbitService, PrismaModule } from '@coldpbc/nest';
+import { CacheService, ColdCacheModule, ColdRabbitModule, ColdRabbitService, MqttModule, PrismaModule } from '@coldpbc/nest';
 import { RabbitService } from './rabbit.service';
 import { ServiceDefinitionsService } from './service_definitions.service';
 import { ServiceDefinitionsController } from './service_definitions.controller';
 
 @Module({
-  imports: [PrismaModule, ColdCacheModule, ColdRabbitModule.forFeature()],
+  imports: [PrismaModule, ColdCacheModule, ColdRabbitModule.forFeature(), MqttModule],
   controllers: [ServiceDefinitionsController],
   providers: [ServiceDefinitionsService, CacheService, ColdRabbitService, RabbitService],
+  exports: [ServiceDefinitionsService, CacheService, ColdRabbitService, RabbitService],
 })
 export class Service_definitionsModule {}