+++ libs/nest/prisma/migrations/20240105213214_add_organization_id/migration.sql
@@ -1,7 +1,7 @@
 /*
   Warnings:
 
-  - You are about to drop the column `integration_id` on the `organization_locations` table. All the data in the column will be lost.
+  - You are about to drop the column `integration_id` on the `locations` table. All the data in the column will be lost.
 
 */
 -- DropForeignKey