+++ libs/nest/prisma/migrations/20240108181203_rename_to_postal_code/migration.sql
@@ -1,7 +1,7 @@
 /*
   Warnings:
 
-  - You are about to drop the column `zip` on the `organization_locations` table. All the data in the column will be lost.
+  - You are about to drop the column `zip` on the `locations` table. All the data in the column will be lost.
 
 */
 -- AlterTable