+++ libs/nest/src/lib/aws/index.ts
@@ -1,1 +1,2 @@
 export * from "./s3/index"
+export * from "./secrets/index"