+++ libs/nest/src/lib/mqtt/mqtt.module.ts
@@ -1,0 +1,9 @@
+import { Module } from '@nestjs/common';
+import { MqttService } from './mqtt.service';
+import { MqttValidatorService } from './validator/mqtt.validator.service';
+
+@Module({
+  providers: [MqttService, MqttValidatorService],
+  exports: [MqttService],
+})
+export class MqttModule {}